RNAseq Differential Expressison

This function allows you to identify differentially expressed genes.

rna_diff_expr(count_table, design_table, method = "DESeq2")

Examples

rna_diff_expr(count_table, design_table, method="DESeq")

#> 
#> Attaching package: 'dplyr'

#> The following objects are masked from 'package:stats':
#> 
#>     filter, lag

#> The following objects are masked from 'package:base':
#> 
#>     intersect, setdiff, setequal, union

#> [1] "using DESeq"

#> Loading required package: BiocGenerics

#> Loading required package: parallel

#> 
#> Attaching package: 'BiocGenerics'

#> The following objects are masked from 'package:parallel':
#> 
#>     clusterApply, clusterApplyLB, clusterCall, clusterEvalQ,
#>     clusterExport, clusterMap, parApply, parCapply, parLapply,
#>     parLapplyLB, parRapply, parSapply, parSapplyLB

#> The following objects are masked from 'package:dplyr':
#> 
#>     combine, intersect, setdiff, union

#> The following objects are masked from 'package:stats':
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from 'package:base':
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colMeans,
#>     colnames, colSums, dirname, do.call, duplicated, eval, evalq,
#>     Filter, Find, get, grep, grepl, intersect, is.unsorted, lapply,
#>     lengths, Map, mapply, match, mget, order, paste, pmax, pmax.int,
#>     pmin, pmin.int, Position, rank, rbind, Reduce, rowMeans, rownames,
#>     rowSums, sapply, setdiff, sort, table, tapply, union, unique,
#>     unsplit, which, which.max, which.min

#> Loading required package: Biobase

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> Loading required package: locfit

#> locfit 1.5-9.1 	 2013-03-22

#> Loading required package: lattice

#>     Welcome to 'DESeq'. For improved performance, usability and
#>     functionality, please consider migrating to 'DESeq2'.

#> Loading required package: S4Vectors

#> Loading required package: stats4

#> 
#> Attaching package: 'S4Vectors'

#> The following objects are masked from 'package:dplyr':
#> 
#>     first, rename

#> The following object is masked from 'package:base':
#> 
#>     expand.grid

#> Loading required package: IRanges

#> 
#> Attaching package: 'IRanges'

#> The following objects are masked from 'package:dplyr':
#> 
#>     collapse, desc, slice

#> The following object is masked from 'package:grDevices':
#> 
#>     windows

#> Loading required package: GenomicRanges

#> Loading required package: GenomeInfoDb

#> Loading required package: SummarizedExperiment

#> Loading required package: DelayedArray

#> Loading required package: matrixStats

#> 
#> Attaching package: 'matrixStats'

#> The following objects are masked from 'package:Biobase':
#> 
#>     anyMissing, rowMedians

#> The following object is masked from 'package:dplyr':
#> 
#>     count

#> Loading required package: BiocParallel

#> 
#> Attaching package: 'DelayedArray'

#> The following objects are masked from 'package:matrixStats':
#> 
#>     colMaxs, colMins, colRanges, rowMaxs, rowMins, rowRanges

#> The following objects are masked from 'package:base':
#> 
#>     aperm, apply

#> 
#> Attaching package: 'DESeq2'

#> The following objects are masked from 'package:DESeq':
#> 
#>     estimateSizeFactorsForMatrix, getVarianceStabilizedData,
#>     varianceStabilizingTransformation

#> Error in ncol(countData) == nrow(colData): object 'count_table' not found

rna_diff_expr(count_table, design_table, method="DESeq2")

#> [1] "using DESeq2"

#> Error in ncol(countData) == nrow(colData): object 'count_table' not found

rna_diff_expr(count_table, design_table, method="edgeR")

#> [1] "using edgeR"

#> Loading required package: limma

#> 
#> Attaching package: 'limma'

#> The following object is masked from 'package:DESeq2':
#> 
#>     plotMA

#> The following object is masked from 'package:DESeq':
#> 
#>     plotMA

#> The following object is masked from 'package:BiocGenerics':
#> 
#>     plotMA

#> Error in ncol(countData) == nrow(colData): object 'count_table' not found

rna_diff_expr(count_table, design_table, method="limma-voom")

#> [1] "using limma-voom"

#> Error in ncol(countData) == nrow(colData): object 'count_table' not found

rna_diff_expr(count_table, design_table, method="sleuth")

#> [1] "using sleuth"

#> [1] 0

rna_diff_expr(count_table, design_table, method="baySeq")

#> [1] "using baySeq"

#> Loading required package: abind

#> 
#> Attaching package: 'baySeq'

#> The following object is masked from 'package:dplyr':
#> 
#>     groups

#> Finding priors...

#> done.

#> Length of priorReps:1000

#> Length of priorSubset:1000

#> Length of subset:1000

#> Length of postRows:1000

#> ...

#>    annotation simA.1 simA.2 simA.3 simA.4 simA.5 simB.1 simB.2 simB.3 simB.4
#> 1    count_80      1      1      0      1      1     13     21      8      6
#> 2    count_78      1      1      0      1      1      8     13      7      9
#> 3    count_66      0      0      0      0      0     15     10      4      4
#> 4    count_21      2      0      1      1      0     15     15      6      5
#> 5     count_7      9      8      8      8      9      1      2      1      0
#> 6    count_26     13      4     11      5      7      1      1      1      0
#> 7    count_64      6      6      8     11      9      1      1      0      0
#> 8    count_72      0      0      1      0      0      7      6      4      3
#> 9    count_83     14      6      9      2      9      1      0      0      1
#> 10   count_27      5      3      6      4      7      0      0      0      1
#>    simB.5 Likelihood ordering       FDR.DE      FWER.DE
#> 1      20  0.9995791      2>1 0.0004208781 0.0004208781
#> 2      10  0.9992829      2>1 0.0005689902 0.0011376785
#> 3      10  0.9990943      2>1 0.0006812318 0.0020423631
#> 4      11  0.9985733      2>1 0.0008675879 0.0034661055
#> 5       0  0.9983926      1>2 0.0010155560 0.0050679624
#> 6       0  0.9980266      1>2 0.0011752048 0.0070314101
#> 7       1  0.9962283      1>2 0.0015461322 0.0107765864
#> 8       8  0.9957691      2>1 0.0018817294 0.0149619013
#> 9       1  0.9954451      1>2 0.0021787487 0.0194486539
#> 10      0  0.9941751      1>2 0.0025433681 0.0251603097

rna_diff_expr(count_table, design_table, method="NOISeq")

#> [1] "using NOISeq"

#> [1] 0

rna_diff_expr(count_table, design_table, method="EBSeq")

#> [1] "using EBSeq"

#> [1] 0